# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CreatorAccess {
  authorized: boolean;
}

export interface FeedbackMessage {
  id: string;
  /** @nullable */
  name: string | null;
  message: string;
  createdAt: string;
}

export interface FeedbackInput {
  /** @maxLength 80 */
  name?: string;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  message: string;
}

export interface LeaderboardEntry {
  id: string;
  playerName: string;
  score: number;
  accuracy: number;
  /**
     * @minimum 1
     * @maximum 3
     */
  difficulty: number;
  tableFrom: number;
  tableTo: number;
  playedAt: string;
  correctCount: number;
  totalQuestions: number;
}

export interface CreateLeaderboardEntryBody {
  runId: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  playerName: string;
  score: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  accuracy: number;
  /**
     * @minimum 1
     * @maximum 3
     */
  difficulty: number;
  /**
     * @minimum 2
     * @maximum 20
     */
  tableFrom: number;
  /**
     * @minimum 2
     * @maximum 20
     */
  tableTo: number;
  /** @minimum 0 */
  correctCount: number;
  /** @minimum 0 */
  totalQuestions: number;
}