# lib/api-zod/src/generated/types/createLeaderboardEntryBody.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateLeaderboardEntryBody {
  runId: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  playerName: string;
  score: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  accuracy: number;
  /**
     * @minimum 1
     * @maximum 3
     */
  difficulty: number;
  /**
     * @minimum 2
     * @maximum 20
     */
  tableFrom: number;
  /**
     * @minimum 2
     * @maximum 20
     */
  tableTo: number;
  /** @minimum 0 */
  correctCount: number;
  /** @minimum 0 */
  totalQuestions: number;
}