# lib/api-zod/src/generated/types/feedbackInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FeedbackInput {
  /** @maxLength 80 */
  name?: string;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  message: string;
}