# lib/api-zod/src/generated/types/feedbackMessage.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FeedbackMessage {
  id: string;
  /** @nullable */
  name: string | null;
  message: string;
  createdAt: string;
}