# lib/api-zod/src/generated/types/leaderboardEntry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LeaderboardEntry {
  id: string;
  playerName: string;
  score: number;
  accuracy: number;
  /**
     * @minimum 1
     * @maximum 3
     */
  difficulty: number;
  tableFrom: number;
  tableTo: number;
  playedAt: string;
  correctCount: number;
  totalQuestions: number;
}